Module bevy_app::prelude

Re-exports

pub use crate::CoreStage;

pub use crate::StartupStage;

Structs

App
Containers of app logic and data

AppBuilder
Configure Apps using the builder pattern

EventReader
Reads events of type T in order and tracks which events have already been read.

EventWriter
Sends events of type T.

Traits

Plugin
A collection of Bevy App logic and configuration

PluginGroup

Derive Macros

DynamicPlugin
Generates a dynamic plugin entry point function for the given Plugin type.