Struct bevy_app::PluginGroupBuilder

pub struct PluginGroupBuilder { /* private fields */ }

Facilitates the creation and configuration of a PluginGroup. Provides a build ordering to ensure that Plugins which produce/require a Resource are built before/after dependent/depending Plugins. Plugins inside the group can be disabled, enabled or reordered.

Implementations

impl PluginGroupBuilder

pub fn start<PG: PluginGroup>() -> Self

Start a new builder for the PluginGroup.

pub fn set<T: Plugin>(self, plugin: T) -> Self

Sets the value of the given Plugin, if it exists.

Panics

Panics if the Plugin does not exist.

pub fn add<T: Plugin>(self, plugin: T) -> Self

Adds the plugin Plugin at the end of this PluginGroupBuilder. If the plugin was already in the group, it is removed from its previous place.

pub fn add_group(self, group: impl PluginGroup) -> Self

Adds a PluginGroup at the end of this PluginGroupBuilder. If the plugin was already in the group, it is removed from its previous place.

pub fn add_before<Target: Plugin, T: Plugin>(self, plugin: T) -> Self

Adds a Plugin in this PluginGroupBuilder before the plugin of type Target. If the plugin was already the group, it is removed from its previous place. There must be a plugin of type Target in the group or it will panic.

pub fn add_after<Target: Plugin, T: Plugin>(self, plugin: T) -> Self

Adds a Plugin in this PluginGroupBuilder after the plugin of type Target. If the plugin was already the group, it is removed from its previous place. There must be a plugin of type Target in the group or it will panic.

pub fn enable<T: Plugin>(self) -> Self

Enables a Plugin.

Plugins within a PluginGroup are enabled by default. This function is used to opt back in to a Plugin after disabling it. If there are no plugins of type T in this group, it will panic.

pub fn disable<T: Plugin>(self) -> Self

Disables a Plugin, preventing it from being added to the App with the rest of the PluginGroup. The disabled Plugin keeps its place in the PluginGroup, so it can still be used for ordering with add_before or add_after, or it can be re-enabled. If there are no plugins of type T in this group, it will panic.

pub fn finish(self, app: &mut App)

Consumes the PluginGroupBuilder and builds the contained Plugins in the order specified.

Panics

Panics if one of the plugin in the group was already added to the application.