Trait bevy_app::Plugin

pub trait Plugin: Downcast + Any + Send + Sync {
    fn build(&self, app: &mut App);

    fn setup(&self, _app: &mut App) { ... }
    fn name(&self) -> &str { ... }
    fn is_unique(&self) -> bool { ... }
}

A collection of Bevy app logic and configuration.

Plugins configure an App. When an App registers a plugin, the plugin’s Plugin::build function is run. By default, a plugin can only be added once to an App.

If the plugin may need to be added twice or more, the function is_unique() should be overridden to return false. Plugins are considered duplicate if they have the same name(). The default name() implementation returns the type name, which means generic plugins with different type parameters will not be considered duplicates.

Required Methods

fn build(&self, app: &mut App)

Configures the App to which this plugin is added.

Provided Methods

fn setup(&self, _app: &mut App)

Runs after all plugins are built, but before the app runner is called. This can be useful if you have some resource that other plugins need during their build step, but after build you want to remove it and send it to another thread.

fn name(&self) -> &str

Configures a name for the Plugin which is primarily used for checking plugin uniqueness and debugging.

fn is_unique(&self) -> bool

If the plugin can be meaningfully instantiated several times in an App, override this method to return false.

Implementations

impl dyn Plugin

pub fn is<__T: Plugin>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: Plugin>(self: Box<Self>) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn’t.

pub fn downcast_rc<__T: Plugin>(self: Rc<Self>) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn’t.

pub fn downcast_ref<__T: Plugin>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_mut<__T: Plugin>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn’t.